파일 구조 변경, 주석 제거
